+++ secrethold.js
@@ -1,9 +1,13 @@
 'use strict';
 
 const Secrethold = require('./lib');
-const ErrorCodes = require('./lib/errors');
+const ErrorCodes = require('./lib/error-codes');
 const CryptoConstants = require('./lib/cryptography/constants');
 
+/**
+ * Lightweight Node.js library designed to keep secrets.
+ * @module Secrethold
+ */
 module.exports = {
   Secrethold,
   ErrorCodes,